consolidating parquet logic to processor class
