+++ koku/masu/test/util/aws/test_common.py
@@ -16,7 +16,6 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 import random
-import uuid
 from datetime import datetime
 from unittest import TestCase
 from unittest.mock import Mock
@@ -33,7 +32,6 @@ from masu.database.aws_report_db_accessor import AWSReportDBAccessor
 from masu.database.provider_db_accessor import ProviderDBAccessor
 from masu.external import AWS_REGIONS
 from masu.external.date_accessor import DateAccessor
-from masu.processor.report_parquet_processor_base import ReportParquetProcessorBase
 from masu.test import MasuTestCase
 from masu.test.external.downloader.aws import fake_arn
 from masu.test.external.downloader.aws import fake_aws_account_id
@@ -368,123 +366,6 @@ class TestAWSUtils(MasuTestCase):
                 mock_s3.side_effect = ClientError({}, "Error")
                 upload = utils.copy_data_to_s3_bucket("request_id", "path", "filename", "data", "manifest_id")
                 self.assertEqual(upload, None)
-
-    def test_get_file_keys_from_s3_with_manifest_id(self):
-        """Test get_file_keys_from_s3_with_manifest_id."""
-        files = utils.get_file_keys_from_s3_with_manifest_id("request_id", "s3_path", "manifest_id")
-        self.assertEqual(files, [])
-
-        with patch("masu.util.aws.common.settings", ENABLE_S3_ARCHIVING=True):
-            with patch("masu.util.aws.common.get_s3_resource") as mock_s3:
-                files = utils.get_file_keys_from_s3_with_manifest_id("request_id", None, "manifest_id")
-                self.assertEqual(files, [])
-
-        with patch("masu.util.aws.common.settings", ENABLE_S3_ARCHIVING=True):
-            with patch("masu.util.aws.common.get_s3_resource") as mock_s3:
-                mock_s3.side_effect = ClientError({}, "Error")
-                files = utils.get_file_keys_from_s3_with_manifest_id("request_id", "s3_path", "manifest_id")
-                self.assertEqual(files, [])
-
-    def test_convert_csv_to_parquet(self):
-        """Test convert_csv_to_parquet."""
-        result = utils.convert_csv_to_parquet(
-            "request_id", None, "s3_parquet_path", "local_path", "manifest_id", "csv_filename"
-        )
-        self.assertFalse(result)
-
-        result = utils.convert_csv_to_parquet(
-            "request_id", "s3_csv_path", "s3_parquet_path", "local_path", "manifest_id", "csv_filename"
-        )
-        self.assertFalse(result)
-
-        with patch("masu.util.aws.common.settings", ENABLE_S3_ARCHIVING=True):
-            with patch("masu.util.aws.common.get_s3_resource") as mock_s3:
-                with patch("masu.util.aws.common.shutil.rmtree"):
-                    with patch("masu.util.aws.common.Path"):
-                        mock_s3.side_effect = ClientError({}, "Error")
-                        result = utils.convert_csv_to_parquet(
-                            "request_id",
-                            "s3_csv_path",
-                            "s3_parquet_path",
-                            "local_path",
-                            "manifest_id",
-                            "csv_filename.csv",
-                        )
-                        self.assertFalse(result)
-
-        with patch("masu.util.aws.common.settings", ENABLE_S3_ARCHIVING=True):
-            with patch("masu.util.aws.common.get_s3_resource"):
-                with patch("masu.util.aws.common.shutil.rmtree"):
-                    with patch("masu.util.aws.common.Path"):
-                        result = utils.convert_csv_to_parquet(
-                            "request_id",
-                            "s3_csv_path",
-                            "s3_parquet_path",
-                            "local_path",
-                            "manifest_id",
-                            "csv_filename.csv.gz",
-                        )
-                        self.assertFalse(result)
-
-        with patch("masu.util.aws.common.settings", ENABLE_S3_ARCHIVING=True):
-            with patch("masu.util.aws.common.get_s3_resource"):
-                with patch("masu.util.aws.common.shutil.rmtree"):
-                    with patch("masu.util.aws.common.Path"):
-                        with patch("masu.util.aws.common.pd"):
-                            with patch("masu.util.aws.common.open") as mock_open:
-                                mock_open.side_effect = ValueError()
-                                result = utils.convert_csv_to_parquet(
-                                    "request_id",
-                                    "s3_csv_path",
-                                    "s3_parquet_path",
-                                    "local_path",
-                                    "manifest_id",
-                                    "csv_filename.csv.gz",
-                                )
-                                self.assertFalse(result)
-
-        with patch("masu.util.aws.common.settings", ENABLE_S3_ARCHIVING=True):
-            with patch("masu.util.aws.common.get_s3_resource"):
-                with patch("masu.util.aws.common.Path"):
-                    with patch("masu.util.aws.common.shutil.rmtree"):
-                        with patch("masu.util.aws.common.pd"):
-                            with patch("masu.util.aws.common.open"):
-                                with patch("masu.util.aws.common.BytesIO"):
-                                    with patch("masu.util.aws.common.copy_data_to_s3_bucket"):
-                                        with patch("masu.util.aws.common.create_parquet_table"):
-                                            result = utils.convert_csv_to_parquet(
-                                                "request_id",
-                                                "s3_csv_path",
-                                                "s3_parquet_path",
-                                                "local_path",
-                                                "manifest_id",
-                                                "csv_filename.csv.gz",
-                                            )
-                                            self.assertTrue(result)
-
-    @patch.object(ReportParquetProcessorBase, "create_table")
-    def test_create_parquet_table(self, mock_create_table):
-        """Test create_parquet_table function."""
-        test_matrix = [
-            {"provider_uuid": str(self.aws_provider_uuid), "expected_create": True},
-            {"provider_uuid": str(self.ocp_provider_uuid), "expected_create": True},
-            {"provider_uuid": str(self.azure_provider_uuid), "expected_create": True},
-            {"provider_uuid": str(uuid.uuid4()), "expected_create": False},
-        ]
-        account = 10001
-        manifest_id = "1"
-        s3_parquet_path = "data/to/parquet"
-        output_file = "local_path/file.parquet"
-        report_type = None
-
-        for test in test_matrix:
-            provider_uuid = test.get("provider_uuid")
-            utils.create_parquet_table(account, provider_uuid, manifest_id, s3_parquet_path, output_file, report_type)
-            if test.get("expected_create"):
-                mock_create_table.assert_called()
-            else:
-                mock_create_table.assert_not_called()
-            mock_create_table.reset_mock()
 
 
 class AwsArnTest(TestCase):
